Improve robustness of application logic.
